HttpAccess review
